src/main/java/com/stan/springboottrading/service/impl/ProductServiceImpl.java
package com.stan.springboottrading.service.impl;

import com.stan.springboottrading.dataobject.ProductInfo;
import com.stan.springboottrading.enums.ProductStatusEnum;
import com.stan.springboottrading.repository.ProductInfoRepository;
import com.stan.springboottrading.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class ProductServiceImpl implements ProductService {

    @Autowired
    private ProductInfoRepository productInfoRepository;

    @Override
    public Optional<ProductInfo> findById(String productId) {
        return productInfoRepository.findById(productId);
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return productInfoRepository.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return productInfoRepository.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return productInfoRepository.save(productInfo);
    }
}
